Pataisyti 2.1 uzduotyje kintamieji

+++ 2uzd/2_1uzduotis.cpp
@@ -2,7 +2,7 @@ using namespace std;
 #include <iostream>
 
 int main() {
-    int v1, v2, v3, v4, v5;
+    double v1, v2, v3, v4, v5;
     double Vidurkis;
     cout << "Iveskite vidurkius mokiniu atskirdami tarpais" << endl;
     cin >> v1 >> v2 >> v3 >> v4 >> v5;
